Reconstruct one shot strategy
