+++ autogpts/autogpt/autogpt/agents/components.py
@@ -2,7 +2,9 @@ from typing import Callable, Generic, Type, TypeVar
 
 #TODO temporary wrapper for single-result component functions
 T = TypeVar("T")
+
 class Single(Generic[T]):
+    """A wrapper for a single result (non-pipeline) of a component function."""
     def __init__(self, value: T):
         self.value = value
 class Component:
