+++ autogpts/autogpt/autogpt/agents/protocols.py
@@ -1,7 +1,6 @@
-from typing import Callable, Generic, Iterator, Protocol, TypeVar, runtime_checkable
+from typing import Iterator, Protocol, runtime_checkable
 
 from autogpt.agents.base import ThoughtProcessOutput
-from autogpt.core.prompting.schema import ChatPrompt
 from autogpt.core.resource.model_providers.schema import (
     AssistantChatMessage,
     ChatMessage,
@@ -10,24 +9,12 @@ from autogpt.models.command import Command
 from autogpt.agents.components import Single
 
 
-
-@runtime_checkable
-class BuildPrompt(Protocol):
-    def build_prompt(
-        self,
-        messages: list[ChatMessage],
-        commands: list[Command],
-        prompt: ChatPrompt,
-    ) -> Single[ChatPrompt]:
-        ...
-
-
 @runtime_checkable
 class MessageProvider(Protocol):
     def get_messages(self) -> Iterator[ChatMessage]:
         ...
 
-
+#TODO kcze process_action
 @runtime_checkable
 class ProposeAction(Protocol):
     def propose_action(self, result: ThoughtProcessOutput) -> None:
@@ -41,26 +28,26 @@ class CommandProvider(Protocol):
 
 
 @runtime_checkable
+class DirectivesProvider(Protocol):
+    def get_contraints(self) -> Iterator[str]:
+        return iter([])
+
+    def get_resources(self) -> Iterator[str]:
+        return iter([])
+
+    def get_best_practices(self) -> Iterator[str]:
+        return iter([])
+
+
+@runtime_checkable
 class ParseResponse(Protocol):
     def parse_response(
-        self, result: ThoughtProcessOutput, llm_response: AssistantChatMessage
+        self, result: ThoughtProcessOutput, response: AssistantChatMessage
     ) -> Single[ThoughtProcessOutput]:
         ...
 
 
-@runtime_checkable
-class GuidelinesProvider(Protocol):
-    def get_contraints(self) -> Iterator[str]:
-        ...
-
-    def get_resources(self) -> Iterator[str]:
-        ...
-
-    def get_best_practices(self) -> Iterator[str]:
-        ...
-
-
-@runtime_checkable
-class OnExecute(Protocol):
-    def on_execute(self) -> None:
-        ...
+# @runtime_checkable
+# class OnExecute(Protocol):
+#     def on_execute(self) -> None:
+#         ...
